refactor: rename functions & routes
